l7lwa

+++ bonus/ft_colore_bonus.c
@@ -6,7 +6,7 @@
 /*   By: oel-qasr <oel-qasr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/24 10:40:03 by oel-qasr          #+#    #+#             */
-/*   Updated: 2024/07/27 16:55:34 by oel-qasr         ###   ########.fr       */
+/*   Updated: 2024/07/29 08:49:02 by oel-qasr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ void	color_back(t_fdf *box, int x)
 		j = 0;
 		while (j <= WIDTH)
 		{
-			my_pixel_put(j, i, get_cr(c1, c2, speed_color(j, i)), box);
+			my_pixel_put((int)j, (int)i, get_cr(c1, c2, speed_color(j, i)), box);
 			j++;
 		}
 		i++;
